requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
solders==0.23.0
solana==0.36.0
python-dotenv==1.1.0
supabase==2.15.0
requests==2.32.3
networkx==3.2.1 
openai==0.28.0
spl-token==0.2.0